📝: add or improve doc comments
