+++ und.go
@@ -26,11 +26,9 @@ var (
 // Und[T] is a type that can express a value (`T`), empty (`null`), or absent (`undefined`).
 // Und[T] is comparable if T is comparable. And it can be copied by assign.
 //
-// When using Und[T] as a struct field,
-// it can be skipped while marshaling if
-//   - the field is `undefined`
-//   - marshaler is "github.com/go-json-experiment/json/jsontext"
-//   - omitzero options is set.
+// Und[T] implements IsZero and can be skippable struct fields when marshaled through appropriate marshalers,
+// e.g. "github.com/go-json-experiment/json/jsontext" with omitzero option set to the field,
+// or "github.com/json-iterator/go" with omitempty option to the field and an appropriate extension.
 //
 // If you need to stick with encoding/json v1, you can use github.com/ngicks/und/sliceund,
 // a slice based version of Und[T] whish is already skppable by v1.
@@ -38,18 +36,21 @@ type Und[T any] struct {
 	opt option.Option[option.Option[T]]
 }
 
+// Defined returns a defined Und[T] whose internal value is t.
 func Defined[T any](t T) Und[T] {
 	return Und[T]{
 		opt: option.Some(option.Some(t)),
 	}
 }
 
+// Null returns a null Und[T].
 func Null[T any]() Und[T] {
 	return Und[T]{
 		opt: option.Some(option.None[T]()),
 	}
 }
 
+// Undefined returns an undefined Und[T].
 func Undefined[T any]() Und[T] {
 	return Und[T]{}
 }
@@ -64,10 +65,14 @@ func FromPointer[T any](v *T) Und[T] {
 	return Defined(*v)
 }
 
+// FromOptions converts opt into an Und[T].
+// opt is retained by the returned value.
 func FromOption[T any](opt option.Option[option.Option[T]]) Und[T] {
 	return Und[T]{opt: opt}
 }
 
+// FromSqlNull converts a valid sql.Null[T] to a defined Und[T]
+// and invalid one into a null Und[].
 func FromSqlNull[T any](v sql.Null[T]) Und[T] {
 	if !v.Valid {
 		return Null[T]()
@@ -75,30 +80,39 @@ func FromSqlNull[T any](v sql.Null[T]) Und[T] {
 	return Defined(v.V)
 }
 
+// IsZero is an alias for IsUndefined.
 func (u Und[T]) IsZero() bool {
 	return u.IsUndefined()
 }
 
+// IsDefined returns true if u is a defined value, otherwise false.
 func (u Und[T]) IsDefined() bool {
 	return u.opt.IsSome() && u.opt.Value().IsSome()
 }
 
+// IsNull returns true if u is a null value, otherwise false.
 func (u Und[T]) IsNull() bool {
 	return u.opt.IsSome() && u.opt.Value().IsNone()
 }
 
+// IsUndefined returns true if u is an undefined value, otherwise false.
 func (u Und[T]) IsUndefined() bool {
 	return u.opt.IsNone()
 }
 
+// Equal implements Equality[Und[T]].
+// Equal panics if T is uncomparable and does not implement Equality[T].
 func (u Und[T]) Equal(other Und[T]) bool {
 	return u.opt.Equal(other.opt)
 }
 
+// Clone clones u.
+// This is only a copy-by-assign unless T implements Cloner[T].
 func (u Und[T]) Clone() Und[T] {
 	return u.Map(func(o option.Option[option.Option[T]]) option.Option[option.Option[T]] { return o.Clone() })
 }
 
+// Value returns an internal value.
 func (u Und[T]) Value() T {
 	if u.IsDefined() {
 		return u.opt.Value().Value()
@@ -107,6 +121,7 @@ func (u Und[T]) Value() T {
 	return zero
 }
 
+// Pointer returns u's internal value as a pointer.
 func (u Und[T]) Pointer() *T {
 	if !u.IsDefined() {
 		return nil
@@ -114,6 +129,7 @@ func (u Und[T]) Pointer() *T {
 	return u.opt.Value().Pointer()
 }
 
+// DoublePointer returns nil if u is undefined, &(*T)(nil) if null, the internal value if defined.
 func (u Und[T]) DoublePointer() **T {
 	switch {
 	case u.IsUndefined():
@@ -128,14 +144,17 @@ func (u Und[T]) DoublePointer() **T {
 	}
 }
 
+// Unwrap returns u's internal value.
 func (u Und[T]) Unwrap() option.Option[option.Option[T]] {
 	return u.opt
 }
 
+// Map returns a new Und[T] whose internal value is u's mapped by f.
 func (u Und[T]) Map(f func(option.Option[option.Option[T]]) option.Option[option.Option[T]]) Und[T] {
 	return Und[T]{opt: f(u.opt)}
 }
 
+// MarshalJSON implements json.Marshaler.
 func (u Und[T]) MarshalJSON() ([]byte, error) {
 	if u.IsUndefined() || u.IsNull() {
 		return []byte(`null`), nil
@@ -143,6 +162,7 @@ func (u Und[T]) MarshalJSON() ([]byte, error) {
 	return json.Marshal(u.opt.Value().Value())
 }
 
+// UnmarshalJSON implements json.Unmarshaler.
 func (u *Und[T]) UnmarshalJSON(data []byte) error {
 	if string(data) == "null" {
 		*u = Null[T]()
@@ -159,6 +179,7 @@ func (u *Und[T]) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
+// MarshalJSONV2 implements jsonv2.MarshalerV2.
 func (u Und[T]) MarshalJSONV2(enc *jsontext.Encoder, opts jsonv2.Options) error {
 	if !u.IsDefined() {
 		return enc.WriteToken(jsontext.Null)
@@ -166,6 +187,7 @@ func (u Und[T]) MarshalJSONV2(enc *jsontext.Encoder, opts jsonv2.Options) error 
 	return jsonv2.MarshalEncode(enc, u.opt.Value().Value(), opts)
 }
 
+// UnmarshalJSONV2 implements jsonv2.UnmarshalerV2.
 func (u *Und[T]) UnmarshalJSONV2(dec *jsontext.Decoder, opts jsonv2.Options) error {
 	if dec.PeekKind() == 'n' {
 		err := dec.SkipValue()
